+++ src/main/java/lagatrix/tools/gui_factory/AjustTextFactory.java
@@ -1,7 +1,7 @@
 package lagatrix.tools.gui_factory;
 
 /**
- * AJust text in dialog.
+ * Ajust text in dialog.
  *
  * @author javierfh03
  * @since 0.3
@@ -15,10 +15,18 @@ public class AjustTextFactory {
      * @return The text formated.
      */
     public static String ajust(String text) {
+        int pointer = 26;
+        
         if (text.length() > 27) {
+            
+            // Avoid cutting words.
+            while (text.charAt(pointer) != ' ') {                
+                pointer--;
+            }
+            
             return String.format("<html>%s<br>%s</html>", 
-                    text.substring(0, 26),
-                        text.substring(26, text.length()));
+                    text.substring(0, pointer),
+                        text.substring(pointer, text.length()));
         }
         
         return text;